chore: Move interfaces.ts out of searchAndSort folder
