+++ src/components/AutoAssignDialog.test.tsx
@@ -1,6 +1,6 @@
 import { render, fireEvent, screen, within } from "@testing-library/react";
 import "@testing-library/jest-dom";
-import { Metadata } from "../searchAndSort/interfaces";
+import { Metadata } from "../interfaces";
 import { AutoAssignDialog, AssignmentCount } from "./AutoAssignDialog";
 
 const updateAssignees = jest.fn(
