+++ src/searchAndSort/SearchFilterCard.tsx
@@ -2,7 +2,7 @@ import { ReactElement } from "react";
 import { makeStyles, Theme } from "@material-ui/core/styles";
 import { List, ListItem, ListItemText, Card } from "@material-ui/core";
 import { Clear } from "@material-ui/icons";
-import { Filter } from "@/searchAndSort/interfaces";
+import { Filter } from "@/interfaces";
 
 const useStyles = makeStyles((theme: Theme) => ({
   card: {
